+++ src/handler.rs
@@ -27,7 +27,7 @@ pub trait IntoResponse<A: Actor, M: ResponseType> {
 }
 
 impl<A, M> IntoResponse<A, M> for ()
-    where A: Actor + Handler<M>, M: ResponseType<Item=(), Error=()>,
+    where A: Actor, M: ResponseType<Item=(), Error=()>,
 {
     fn into_response(self) -> Response<A, M> {
         Response::reply(Ok(()))
@@ -35,7 +35,7 @@ impl<A, M> IntoResponse<A, M> for ()
 }
 
 impl<A, M> IntoResponse<A, M> for Response<A, M>
-    where A: Actor + Handler<M>, M: ResponseType,
+    where A: Actor, M: ResponseType,
 {
     fn into_response(self) -> Response<A, M> {
         self
@@ -43,7 +43,7 @@ impl<A, M> IntoResponse<A, M> for Response<A, M>
 }
 
 impl<A, M> IntoResponse<A, M> for Result<M::Item, M::Error>
-    where A: Actor + Handler<M>, M: ResponseType,
+    where A: Actor, M: ResponseType,
 {
     fn into_response(self) -> Response<A, M> {
         Response {inner: Some(ResponseTypeItem::Result(self))}
@@ -51,7 +51,7 @@ impl<A, M> IntoResponse<A, M> for Result<M::Item, M::Error>
 }
 
 impl<A, M, B> IntoResponse<A, M> for SyncAddress<B>
-    where A: Actor + Handler<M>, M: ResponseType<Item=SyncAddress<B>, Error=()>,
+    where A: Actor, M: ResponseType<Item=SyncAddress<B>, Error=()>,
           B: Actor<Context=Context<B>>
 {
     fn into_response(self) -> Response<A, M> {
@@ -89,10 +89,15 @@ pub trait Handler<M> where Self: Actor, M: ResponseType
 ///
 /// `StreamHandler` is an extension of a `Handler` with stream specific methods.
 #[allow(unused_variables)]
-pub trait StreamHandler<M, E>: Handler<M> where Self: Actor, M: ResponseType,
+pub trait StreamHandler<M, E> where Self: Actor, M: ResponseType,
 {
+    type Result: IntoResponse<Self, M>;
+
     /// Method is called when stream get polled first time.
     fn started(&mut self, ctx: &mut Self::Context, handle: SpawnHandle) {}
+
+    /// Method is called for every message received by this Actor
+    fn handle(&mut self, msg: M, ctx: &mut Self::Context) -> Self::Result;
 
     /// Method is called when stream finishes.
     ///